chore: move inputs into inputs dir
