+++ src/components/inputs/checkbox/checkbox.test.tsx
